scripts/sorttable: Move code from sorttable.h into sorttable.c
jira KERNEL-1156
Rebuild_History Non-Buildable kernel-6.12.0-211.22.1.el10_2
commit-author Steven Rostedt <rostedt@goodmis.org>
commit 58d87678a0f46c6120904b4326aaf5ebf4454c69

Instead of having the main code live in a header file and included twice
with MACROs that define the Elf structures for 64 bit or 32 bit, move the
code in the C file now that the Elf structures are defined in a union that
has both. All accesses to the Elf structure fields are done through helper
function pointers. If the file being parsed if for a 64 bit architecture,
all the helper functions point to the 64 bit versions to retrieve the Elf
fields. The same is true if the architecture is 32 bit, where the function
pointers will point to the 32 bit helper functions.

Note, when the value of a field can be either 32 bit or 64 bit, a 64 bit
is always returned, as it works for the 32 bit code as well.

This makes the code easier to read and maintain, and it now all exists in
sorttable.c and sorttable.h may be removed.

	Cc: bpf <bpf@vger.kernel.org>
	Cc: Masami Hiramatsu <mhiramat@kernel.org>
	Cc: Mark Rutland <mark.rutland@arm.com>
	Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
	Cc: Andrew Morton <akpm@linux-foundation.org>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Linus Torvalds <torvalds@linux-foundation.org>
	Cc: Masahiro Yamada <masahiroy@kernel.org>
	Cc: Nathan Chancellor <nathan@kernel.org>
	Cc: Nicolas Schier <nicolas@fjasle.eu>
	Cc: Zheng Yejian <zhengyejian1@huawei.com>
	Cc: Martin Kelly <martin.kelly@crowdstrike.com>
	Cc: Christophe Leroy <christophe.leroy@csgroup.eu>
	Cc: Josh Poimboeuf <jpoimboe@redhat.com>
	Cc: Stephen Rothwell <sfr@canb.auug.org.au>
Link: https://lore.kernel.org/20250107223217.6f7f96a5@gandalf.local.home
(cherry picked from commit 58d87678a0f46c6120904b4326aaf5ebf4454c69)
